backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
email-validator
python-multipart
psycopg2-binary
passlib[argon2]
python-dotenv
python-socketio
aiofiles